Added Main Menu Relative Layouts.
changed activity_main_menu.xml, added 2 relative layouts. one is
visible, the other is not visible. The not visible one is a menu that
gets the users number of holes and number of players. These layouts will
be toggled by the new game button, on click of the submit button, we can
call a function to start a new activity.